Refactor composite tests
